+++ lib/guard/test.rb
@@ -7,8 +7,8 @@ require 'test/unit/version'
 module Guard
   class Test < Guard
 
-    autoload :Runner,    'guard/test/runner'
-    autoload :Inspector, 'guard/test/inspector'
+    require 'guard/test/runner'
+    require 'guard/test/inspector'
 
     def initialize(watchers = [], options = {})
       super
@@ -25,7 +25,8 @@ module Guard
     end
 
     def start
-      ::Guard::UI.info("Guard::Test #{TestVersion::VERSION} is running, with Test::Unit #{::Test::Unit::VERSION}!", :reset => true)
+      ::Guard::UI.info("Guard::Test #{TestVersion::VERSION} is running, with Test::Unit #{::Test::Unit::VERSION}!",
+                       :reset => true)
       run_all if @options[:all_on_start]
     end